Move compactor to elev mod
