+++ hexit/src/elevation.rs
@@ -1,5 +1,6 @@
 use anyhow::Result;
 use byteorder::{LittleEndian as LE, ReadBytesExt};
+use hextree::{compaction::Compactor, Cell};
 use std::io::Read;
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
@@ -77,5 +78,28 @@ impl Elevation {
             max: new_max,
             n: new_n,
         })
+    }
+}
+
+pub struct ReductionCompactor {
+    pub target_resolution: u8,
+    pub source_resolution: u8,
+}
+
+impl Compactor<Elevation> for ReductionCompactor {
+    fn compact(&mut self, cell: Cell, children: [Option<&Elevation>; 7]) -> Option<Elevation> {
+        if cell.res() < self.target_resolution {
+            None
+        } else if let [Some(v0), Some(v1), Some(v2), Some(v3), Some(v4), Some(v5), Some(v6)] =
+            children
+        {
+            Some(Elevation::concat(
+                self.source_resolution,
+                cell.res(),
+                &[*v0, *v1, *v2, *v3, *v4, *v5, *v6],
+            ))
+        } else {
+            None
+        }
     }
 }
